% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gpt_bigcode_from_pretrained}
\alias{gpt_bigcode_from_pretrained}
\title{Loads GPTBigCode from HF Model Hub}
\usage{
gpt_bigcode_from_pretrained(identifier, ..., revision = "main")
}
\arguments{
\item{identifier}{The HuggingFace repository to download the model from.}

\item{...}{Additional configuration options.}

\item{revision}{The repository revision to download from. Either a branch name
or a commit hash.}
}
\description{
Loads GPTBigCode from HF Model Hub
}
