% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptbigcode.R
\name{GPTBigCodeAttention}
\alias{GPTBigCodeAttention}
\alias{GPTBigCodeMLP}
\alias{GPTBigCodePreTrainedModel}
\alias{GPTBigCodeModel}
\alias{GPTBigCodeForCausalLM}
\title{GPTBigCode neural net modules}
\usage{
GPTBigCodeAttention(config, is_cross_attention = FALSE, layer_idx = NULL)

GPTBigCodeMLP(intermediate_size, config)

GPTBigCodePreTrainedModel(config)

GPTBigCodeModel(config)

GPTBigCodeForCausalLM(config)
}
\arguments{
\item{config}{Configuration options as created by \code{\link[=gpt_bigcode_config]{gpt_bigcode_config()}}.}

\item{is_cross_attention}{Is cross attention?}

\item{layer_idx}{The layers idx (used to automatically define its size).}

\item{intermediate_size}{the intermediate size}
}
\description{
GPTBigCode neural net modules
}
\section{Functions}{
\itemize{
\item \code{GPTBigCodeMLP()}: MLP module

\item \code{GPTBigCodePreTrainedModel()}: The pretrained model class

\item \code{GPTBigCodeModel()}: The model class

\item \code{GPTBigCodeForCausalLM()}: The text generation class

}}
